+++ native/src/main/java/io/ballerina/lib/avro/deserialize/visitor/RecordUtils.java
@@ -42,7 +42,7 @@ public class RecordUtils {
 
     public static void processMapField(BMap<BString, Object> avroRecord,
                                  Schema.Field field, Object fieldData) throws AvroDeserializationException {
-        Type mapType = extractMapType(avroRecord.getType());
+        Type mapType = extractMapType(avroRecord.getType(), field.name());
         MapDeserializer mapDeserializer = new MapDeserializer(field.schema(), mapType);
         Object fieldValue = mapDeserializer.accept(new DeserializeVisitor(), fieldData);
         avroRecord.put(fromString(field.name()), fieldValue);
@@ -64,7 +64,7 @@ public class RecordUtils {
 
     public static void processRecordField(BMap<BString, Object> avroRecord,
                                     Schema.Field field, Object fieldData) throws AvroDeserializationException {
-        Type recType = extractRecordType((RecordType) avroRecord.getType());
+        Type recType = extractRecordType((RecordType) avroRecord.getType(), field.name());
         RecordDeserializer recordDes = new RecordDeserializer(recType, field.schema());
         Object fieldValue = recordDes.accept(new DeserializeVisitor(), fieldData);
         avroRecord.put(fromString(field.name()), fieldValue);